[storage] Integrate create and drop table with metadata store (#648)
* rename table id to rowstore table id

* create table and drop table

* lint

* add backend test

* test metadata store deletion

* check affected rows

* rename to row store lowercase

* rename to mooncake table id

* unify table id conceots

* code cleanup

+++ src/moonlink/src/storage/storage_utils.rs
@@ -63,7 +63,7 @@ pub struct FileId(pub(crate) u64);
 
 /// Unique table id.
 #[derive(Debug, Clone, PartialEq, Eq, Copy, Hash)]
-pub struct TableId(pub(crate) u64);
+pub struct TableId(pub(crate) u32);
 
 /// A globally unique id for a file.
 #[derive(Debug, Clone, PartialEq, Eq, Copy)]
